extract common types
