+++ typings/createAggregate.d.ts
@@ -1,6 +1,11 @@
-import { Reducer } from 'redux'
-import { KeyMap, R, A1, A2, DiffKey, HasKeysDiff } from './utils'
-import { ActionType, ActionTypes } from './commons'
+import { KeyMap, A1, A2 } from './utils'
+import {
+  ActionType,
+  ActionTypes,
+  ReducerFactory,
+  ActionProvider,
+  Subscriptions
+} from './commons'
 import { Actions, ActionsSrc } from './createActions'
 
 // ______________________________________________________
@@ -17,28 +22,6 @@ type ActionCreators<T> = { readonly [K in keyof T]: ActionCreator<T[K]> }
 
 // ______________________________________________________
 
-type HasKeysDiffErrorMessage = 'SUBSCRIPTIONS_KEY_NOT_MATCH'
-type PayloadErrorMessage = 'PAYLOAD_SCHEMA_NOT_MATCH'
-type ReducerFactory = <S>(state: S) => Reducer<S>
-type SubscribeActions<T, M> = R<T> extends A2<M> ? M : PayloadErrorMessage
-type SubscribeAggregate<T, M> = A2<M> extends A2<T> ? M : PayloadErrorMessage
-type SubscribeActionsMap<T, M> = HasKeysDiff<T, M> extends false
-  ? { [K in keyof T & keyof M]?: SubscribeActions<T[K], M[K]> } & KeyMap
-  : HasKeysDiffErrorMessage
-type SubscribeAggregateMap<T, M> = HasKeysDiff<T, M> extends false
-  ? { [K in keyof T & keyof M]?: SubscribeAggregate<T[K], M[K]> } & KeyMap
-  : HasKeysDiffErrorMessage
-
-// ______________________________________________________
-
-type ISM = { __srcmap__: any }
-type SrcMap<T extends ISM> = T['__srcmap__']
-type ActionProvider<T extends ISM> =
-  | Aggregate<SrcMap<T>>
-  | Actions<SrcMap<T>>
-type Subscriptions<T extends ISM, M> = T extends Aggregate<SrcMap<T>>
-  ? SubscribeAggregateMap<SrcMap<T>, M>
-  : SubscribeActionsMap<SrcMap<T>, M>
 interface Aggregate<T> {
   readonly __namespace__: string
   readonly __srcmap__: T
@@ -58,15 +41,4 @@ declare function createAggregate<T extends KeyMap & Mutations<T>>(
   namespace: string
 ): Aggregate<T>
 
-type Modeler<T> = (injects?: Partial<T>) => T
-
-export {
-  Mutations,
-  ActionCreators,
-  ReducerFactory,
-  Aggregate,
-  ActionProvider,
-  Subscriptions,
-  createAggregate,
-  Modeler
-}
+export { Mutations, ActionCreators, Aggregate, createAggregate }
